Move API comments check and AST helper functions to their own files
